ai/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
groq>=0.12.0
sentence-transformers>=3.0.0
weaviate-client>=4.9.0
unstructured[pdf,docx]>=0.16.0
httpx>=0.27.0,<0.28
boto3>=1.35.0
tiktoken>=0.7.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
